% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzohobooks-package.R
\docType{package}
\name{rzohobooks-package}
\alias{rzohobooks}
\alias{rzohobooks-package}
\title{rzohobooks: Work with 'Zoho Books API'}
\description{
Get data from 'Zoho Books API'.
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
}

}
\keyword{internal}
