% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_bank_accounts.R
\name{zb_get_bank_accounts}
\alias{zb_get_bank_accounts}
\title{Get bank accounts}
\usage{
zb_get_bank_accounts(organization_id)
}
\arguments{
\item{organization_id}{Organizations ids}
}
\value{
tibble
}
\description{
Get bank accounts
}
\details{
See \href{https://www.zoho.com/books/api/v3/bank-accounts/#overview}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
bank_accounts <- zb_get_bank_accounts(organizations$organization_id)
}
}
