% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_bank_accounts_details.R
\name{zb_get_bank_accounts_details}
\alias{zb_get_bank_accounts_details}
\title{Get bank accounts details}
\usage{
zb_get_bank_accounts_details(organization_id, account_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{account_id}{Vector of account ids}
}
\value{
tibble
}
\description{
Get bank accounts details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/bankaccounts/#get-a-bank-account
}
\examples{
\dontrun{
bank_accounts <- zb_get_bank_accounts(organizations$organization_id)
details <- zb_get_bank_accounts_details(
  organization_id = organizations$organization_id[1],
  account_id      = bank_accounts$account_id
)
}
}
