% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_bank_transactions.R
\name{zb_get_bank_transactions}
\alias{zb_get_bank_transactions}
\title{Get list of bank transactions}
\usage{
zb_get_bank_transactions(
  organization_id,
  date_start = NULL,
  date_end = NULL,
  search_text = NULL,
  status = c("All", "uncategorized", "manually_added", "matched", "excluded",
    "categorized")
)
}
\arguments{
\item{organization_id}{Organizations ids}

\item{date_start}{Start date, to provide a range within which the transaction date exist}

\item{date_end}{End date, to provide a range within which the transaction date exist}

\item{search_text}{Search Transactions by contact name or description}

\item{status}{Transaction status wise list view - All, uncategorized, manually_added, matched, excluded, categorized}
}
\value{
tibble
}
\description{
Get list of bank transactions
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
bank_transactions <- zb_get_bank_transactions(
    organizations$organization_id,
    date_start = '2025-02-01',
    date_end = '2025-02-10'
)
}
}
