% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_bank_transactions_details.R
\name{zb_get_bank_transactions_details}
\alias{zb_get_bank_transactions_details}
\title{Get bank transactions details}
\usage{
zb_get_bank_transactions_details(organization_id, transaction_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{transaction_id}{Vector of transaction ids}
}
\value{
tibble
}
\description{
Get bank transactions details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/banktransactions/#get-a-transaction
}
\examples{
\dontrun{
bank_transactions <- zb_get_bank_transactions(organizations$organization_id)
details <- zb_get_bank_transactions_details(
  organization_id = organizations$organization_id[1],
  transaction_id  = bank_transactions$transaction_id
)
}
}
