% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_bills.R
\name{zb_get_bills}
\alias{zb_get_bills}
\title{Get list of bills}
\usage{
zb_get_bills(
  organization_id,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL,
  search_text = NULL,
  status = c("paid", "open", "overdue", "void and partially_paid")
)
}
\arguments{
\item{organization_id}{Organizations id, you can set only one id}

\item{date_start}{Search bills by bill date.}

\item{date_end}{Search bills by bill date.}

\item{date_before}{Search bills by bill date.}

\item{date_after}{Search bills by bill date.}

\item{search_text}{Search bills by bill number or reference number or vendor name.}

\item{status}{Search bills by bill status. Allowed Values: paid, open, overdue, void and partially_paid. NULL if you need all bills.}
}
\value{
tibble
}
\description{
Get list of bills
}
\details{
For more detail see \href{https://www.zoho.com/books/api/v3/bills/#list-bills}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
bills <- zb_get_bills(
    organizations$organization_id,
    date_start = '2025-02-01',
    date_end = '2025-02-10'
)
}
}
