% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_bills_details.R
\name{zb_get_bills_details}
\alias{zb_get_bills_details}
\title{Get bills details}
\usage{
zb_get_bills_details(organization_id, bill_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{bill_id}{Vector of bill ids}
}
\value{
tibble
}
\description{
Get bills details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/bills/#get-a-bill
}
\examples{
\dontrun{
bills <- zb_get_bills(organizations$organization_id)
details <- zb_get_bills_details(
  organization_id = organizations$organization_id[1],
  bill_id         = bills$bill_id
)
}
}
