% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_chart_of_accounts.R
\name{zb_get_chart_of_accounts}
\alias{zb_get_chart_of_accounts}
\title{Get List Chart of Accounts}
\usage{
zb_get_chart_of_accounts(organization_id, filter_by = NULL)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter accounts based on its account type and status. Allowed Values: \code{AccountType.All}, \code{AccountType.Active}, \code{AccountType.Inactive}, \code{AccountType.Asset}, \code{AccountType.Liability}, \code{AccountType.Equity}, \code{AccountType.Income} and \code{AccountType.Expense}.}
}
\value{
tibble
}
\description{
Get List Chart of Accounts
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
chart_of_accounts <- zb_get_chart_of_accounts(
    organizations$organization_id
)
}
}
