% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_contacts.R
\name{zb_get_contacts}
\alias{zb_get_contacts}
\title{Get contacts}
\usage{
zb_get_contacts(organization_id)
}
\arguments{
\item{organization_id}{Organizations ids}
}
\value{
tibble
}
\description{
Get contacts
}
\details{
See \href{https://www.zoho.com/books/api/v3/contacts/#list-contacts}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
contacts <- zb_get_contacts(organizations$organization_id)
}
}
