% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_credit_notes_details.R
\name{zb_get_credit_notes_details}
\alias{zb_get_credit_notes_details}
\title{Get credit notes details}
\usage{
zb_get_credit_notes_details(organization_id, creditnote_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{creditnote_id}{Vector of credit note ids}
}
\value{
tibble
}
\description{
Get credit notes details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/credit-notes/#get-a-credit-note
}
\examples{
\dontrun{
notes <- zb_get_credit_notes(organizations$organization_id)
details <- zb_get_credit_notes_details(
  organization_id = organizations$organization_id[1],
  creditnote_id   = notes$creditnote_id
)
}
}
