% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_currencies.R
\name{zb_get_currencies}
\alias{zb_get_currencies}
\title{Get currencies}
\usage{
zb_get_currencies(organization_id)
}
\arguments{
\item{organization_id}{Organizations ids}
}
\value{
tibble
}
\description{
Get currencies
}
\details{
See \href{https://www.zoho.com/books/api/v3/currency/#list-currencies}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
currencies <- zb_get_currencies(organizations$organization_id)
}
}
