% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_customer_payments.R
\name{zb_get_customer_payments}
\alias{zb_get_customer_payments}
\title{Get Payments Received}
\usage{
zb_get_customer_payments(
  organization_id,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL,
  search_text = NULL,
  filter_by = NULL
)
}
\arguments{
\item{organization_id}{Organizations ids}

\item{date_start}{Search payments by estimate date.}

\item{date_end}{Search payments by estimate date.}

\item{date_before}{Search payments by estimate date.}

\item{date_after}{Search payments by estimate date.}

\item{search_text}{Search payments by reference number or customer name or payment description.}

\item{filter_by}{Filter payments by mode.Allowed Values: PaymentMode.All, PaymentMode.Check, PaymentMode.Cash, PaymentMode.BankTransfer, PaymentMode.Paypal, PaymentMode.CreditCard, PaymentMode.GoogleCheckout, PaymentMode.Credit, PaymentMode.Authorizenet, PaymentMode.BankRemittance, PaymentMode.Payflowpro, PaymentMode.Stripe, PaymentMode.TwoCheckout, PaymentMode.Braintree and PaymentMode.Others}
}
\value{
tibble
}
\description{
Get Payments Received
}
\details{
For more infomation see \href{https://www.zoho.com/books/api/v3/customer-payments/#list-customer-payments}{api docs}
}
