% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_customer_payments_details.R
\name{zb_get_customer_payments_details}
\alias{zb_get_customer_payments_details}
\title{Get customer payments details}
\usage{
zb_get_customer_payments_details(organization_id, payment_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{payment_id}{Vector of customer payment ids}
}
\value{
tibble
}
\description{
Get customer payments details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/customerpayments/#get-a-customer-payment
}
\examples{
\dontrun{
payments <- zb_get_customer_payments(organizations$organization_id)
details <- zb_get_customer_payments_details(
  organization_id = organizations$organization_id[1],
  payment_id      = payments$payment_id
)
}
}
