% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_estimates_details.R
\name{zb_get_estimates_details}
\alias{zb_get_estimates_details}
\title{Get estimates details}
\usage{
zb_get_estimates_details(organization_id, estimate_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{estimate_id}{Vector of estimate ids}
}
\value{
tibble
}
\description{
Get estimates details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/estimates/#get-an-estimate
}
\examples{
\dontrun{
estimates <- zb_get_estimates(organizations$organization_id)
details <- zb_get_estimates_details(
  organization_id = organizations$organization_id[1],
  estimate_id     = estimates$estimate_id
)
}
}
