% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_exchange_rates.R
\name{zb_get_exchange_rates}
\alias{zb_get_exchange_rates}
\title{Get currency exchange rates}
\usage{
zb_get_exchange_rates(
  organization_id,
  currency_id,
  from_date = NULL,
  is_current_date = NULL
)
}
\arguments{
\item{organization_id}{Organizations id, you can set only one id}

\item{currency_id}{Id of currency, use \code{zb_get_currencies()} to takes currency ids}

\item{from_date}{Returns the exchange rate details from the given date or from previous closest match in the absence of the exchange rate on the given date.}

\item{is_current_date}{To return the exchange rate only if available for current date.}
}
\value{
tibble
}
\description{
Get currency exchange rates
}
\details{
See \href{https://www.zoho.com/books/api/v3/currency/#list-exchange-rates}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
cur <- zb_get_currencies(organizations$organization_id)
cur_id <- filter(cur, organization_id == organizations$organization_id[1])$currency_id
exc <- zb_get_exchange_rates(organization_id = '20103522465', currency_id = cur_id[11])
}
}
