% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_expenses.R
\name{zb_get_expenses}
\alias{zb_get_expenses}
\title{Get List of All Expenses}
\usage{
zb_get_expenses(
  organization_id,
  filter_by = NULL,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL
)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter expenses by status, allowed values:  \code{Status.All}, \code{Status.Billable}, \code{Status.Nonbillable}, \code{Status.Reimbursed}, \code{Status.Invoiced} \code{Status.Unbilled}}

\item{date_start}{Start date of period}

\item{date_end}{End date of period}

\item{date_before}{Date before}

\item{date_after}{Date after}
}
\value{
tibble
}
\description{
Get List of All Expenses
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
projects <- zb_get_expenses(
    organizations$organization_id
)
}
}
