% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_expenses_details.R
\name{zb_get_expenses_details}
\alias{zb_get_expenses_details}
\title{Get expenses details}
\usage{
zb_get_expenses_details(organization_id, expense_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{expense_id}{Vector of expense ids}
}
\value{
tibble
}
\description{
Get expenses details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/expenses/#get-an-expense
}
\examples{
\dontrun{
expenses <- zb_get_expenses(organizations$organization_id)
details <- zb_get_expenses_details(
  organization_id = organizations$organization_id[1],
  expense_id      = expenses$expense_id
)
}
}
