% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_invoices.R
\name{zb_get_invoices}
\alias{zb_get_invoices}
\title{Get list of invoices}
\usage{
zb_get_invoices(
  organization_id,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL,
  due_date_start = NULL,
  due_date_end = NULL,
  due_date_before = NULL,
  due_date_after = NULL,
  status = NULL,
  search_text = NULL,
  filter_by = NULL
)
}
\arguments{
\item{organization_id}{Organizations ids}

\item{date_start}{Search invoices by invoice date. Default date format is yyyy-mm-dd.}

\item{date_end}{Search invoices by invoice date. Default date format is yyyy-mm-dd.}

\item{date_before}{Search invoices by invoice date. Default date format is yyyy-mm-dd.}

\item{date_after}{Search invoices by invoice date. Default date format is yyyy-mm-dd.}

\item{due_date_start}{Search invoices by due date. Default date format is yyyy-mm-dd.}

\item{due_date_end}{Search invoices by due date. Default date format is yyyy-mm-dd.}

\item{due_date_before}{Search invoices by due date. Default date format is yyyy-mm-dd.}

\item{due_date_after}{Search invoices by due date. Default date format is yyyy-mm-dd.}

\item{status}{Search invoices by invoice status.Allowed Values: sent, draft, overdue, paid, void, unpaid, partially_paid and viewed}

\item{search_text}{Search invoices by invoice number or purchase order or customer name. Max-length 100}

\item{filter_by}{Filter invoices by any status or payment expected date.Allowed Values: \code{Status.All}, \code{Status.Sent}, \code{Status.Draft}, \code{Status.OverDue}, \code{Status.Paid}, \code{Status.Void}, \code{Status.Unpaid}, \code{Status.PartiallyPaid}, \code{Status.Viewed} and \code{Date.PaymentExpectedDate}}
}
\value{
tibble
}
\description{
Get list of invoices
}
\details{
For more details see \href{https://www.zoho.com/books/api/v3/invoices/#list-invoices}{API docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
invoices <- zb_get_invoices(
    organizations$organization_id,
    date_start = '2025-02-01',
    date_end  = '2025-02-10'
 )
}
}
