% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_invoices_details.R
\name{zb_get_invoices_details}
\alias{zb_get_invoices_details}
\title{Get invoices details}
\usage{
zb_get_invoices_details(organization_id, invoice_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{invoice_id}{Vector of invoice ids}
}
\value{
tibble
}
\description{
Get invoices details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/invoices/#get-an-invoice
}
\examples{
\dontrun{
invoices <- zb_get_invoices(organizations$organization_id)
details <- zb_get_invoices_details(
  organization_id = organizations$organization_id[1],
  invoice_id      = invoices$invoice_id
)
}
}
