% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_items.R
\name{zb_get_items}
\alias{zb_get_items}
\title{Get List of All Items}
\usage{
zb_get_items(organization_id, filter_by = NULL)
}
\arguments{
\item{organization_id}{Organizations id, you can set only one id}

\item{filter_by}{Filter items by status, allowed values: \code{Status.All}, \code{Status.Active} and \code{Status.Inactive}}
}
\value{
tibble
}
\description{
Get List of All Items
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
items <- zb_get_items(
    organizations$organization_id
)
}
}
