% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_items_details.R
\name{zb_get_items_details}
\alias{zb_get_items_details}
\title{Get items details}
\usage{
zb_get_items_details(organization_id, item_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{item_id}{Vector of item ids}
}
\value{
tibble
}
\description{
Get items details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/items/#get-an-item
}
\examples{
\dontrun{
items <- zb_get_items(organizations$organization_id)
details <- zb_get_items_details(
  organization_id = organizations$organization_id[1],
  item_id         = items$item_id
)
}
}
