% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_journals.R
\name{zb_get_journals}
\alias{zb_get_journals}
\title{Get Journal List}
\usage{
zb_get_journals(
  organization_id,
  filter_by = NULL,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL
)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter journals by journal date. Allowed Values: \code{JournalDate.All}, \code{JournalDate.Today}, \code{JournalDate.ThisWeek}, \code{JournalDate.ThisMonth}, \code{JournalDate.ThisQuarter} and \code{JournalDate.ThisYear}.}

\item{date_start}{Start date of period}

\item{date_end}{End date of period}

\item{date_before}{Date before}

\item{date_after}{Date after}
}
\value{
tibble
}
\description{
Get Journal List
}
\examples{
\dontrun{
organizations <- zb_get_journals()
projects <- zb_get_expenses(
    organizations$organization_id,
    date_end = '2025-12-01', date_end = '2025-12-15'
)
}
}
