% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_journals_details.R
\name{zb_get_journals_details}
\alias{zb_get_journals_details}
\title{Get journals details}
\usage{
zb_get_journals_details(organization_id, journal_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{journal_id}{Vector of journal ids}
}
\value{
tibble
}
\description{
Get journals details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/journals/#get-a-journal
}
\examples{
\dontrun{
journals <- zb_get_journals(organizations$organization_id)
details <- zb_get_journals_details(
  organization_id = organizations$organization_id[1],
  journal_id      = journals$journal_id
)
}
}
