% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_organizations.R
\name{zb_get_organizations}
\alias{zb_get_organizations}
\title{Get list of organizations}
\usage{
zb_get_organizations()
}
\value{
tibble
}
\description{
Get list of organizations
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
}
}
