% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_projects.R
\name{zb_get_projects}
\alias{zb_get_projects}
\title{Get List of All Projects}
\usage{
zb_get_projects(organization_id, filter_by = NULL)
}
\arguments{
\item{organization_id}{Organizations id, you can set only one id}

\item{filter_by}{Filter projects by status, allowed values: \code{Status.All}, \code{Status.Active} and \code{Status.Inactive}}
}
\value{
tibble
}
\description{
Get List of All Projects
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
projects <- zb_get_projects(
    organizations$organization_id
)
}
}
