% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_recurring_invoice.R
\name{zb_get_recurring_invoice}
\alias{zb_get_recurring_invoice}
\title{Get List of Recurring Invoice}
\usage{
zb_get_recurring_invoice(organization_id, filter_by = NULL)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter Recurring invoices by any status or payment expected date. Allowed Values: \code{Status.All}, \code{Status.Active}, \code{Status.Stopped}, \code{Status.Expired}.}
}
\value{
tibble
}
\description{
Get List of Recurring Invoice
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
recurring_invoice <- zb_get_recurring_invoice(
    organizations$organization_id
)
}
}
