% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_recurring_invoice_details.R
\name{zb_get_recurring_invoice_details}
\alias{zb_get_recurring_invoice_details}
\title{Get recurring invoice details}
\usage{
zb_get_recurring_invoice_details(organization_id, recurring_invoice_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{recurring_invoice_id}{Vector of recurring invoice ids}
}
\value{
tibble
}
\description{
Get recurring invoice details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/recurringinvoices/#get-a-recurring-invoice
}
\examples{
\dontrun{
recurring_invoices <- zb_get_recurring_invoice(organizations$organization_id)
details <- zb_get_recurring_invoice_details(
  organization_id      = organizations$organization_id[1],
  recurring_invoice_id = recurring_invoices$recurring_invoice_id
)
}
}
