% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_retainer_invoices_details.R
\name{zb_get_retainer_invoices_details}
\alias{zb_get_retainer_invoices_details}
\title{Get retainer invoices details}
\usage{
zb_get_retainer_invoices_details(organization_id, retainer_invoice_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{retainer_invoice_id}{Vector of retainer invoice ids}
}
\value{
tibble
}
\description{
Get retainer invoices details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/retainerinvoices/#get-a-retainer-invoice
}
\examples{
\dontrun{
retainer_invoices <- zb_get_retainer_invoices(organizations$organization_id)
details <- zb_get_retainer_invoices_details(
  organization_id     = organizations$organization_id[1],
  retainer_invoice_id = retainer_invoices$retainer_invoice_id
)
}
}
