% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_taxes.R
\name{zb_get_taxes}
\alias{zb_get_taxes}
\title{Get list of taxes}
\usage{
zb_get_taxes(organization_id)
}
\arguments{
\item{organization_id}{Organizations ids}
}
\value{
tibble
}
\description{
Get list of taxes
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
taxes <- zb_get_taxes(organizations$organization_id)
}
}
