% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_users.R
\name{zb_get_users}
\alias{zb_get_users}
\title{Get users list}
\usage{
zb_get_users(organization_id)
}
\arguments{
\item{organization_id}{Organizations ids}
}
\value{
tibble
}
\description{
Get users list
}
\details{
See \href{https://www.zoho.com/books/api/v3/users/#list-users}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
users <- zb_get_users(organizations$organization_id)
}
}
