% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_vendor_credits.R
\name{zb_get_vendor_credits}
\alias{zb_get_vendor_credits}
\title{Get List of Vendor Credits}
\usage{
zb_get_vendor_credits(organization_id, filter_by = NULL)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter vendor credits by status using predefined status values. Allowed values: \code{Status.All}, \code{Status.Open}, \code{Status.Draft}, \code{Status.Closed}, and \code{Status.Void}.}
}
\value{
tibble
}
\description{
Get List of Vendor Credits
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
vendor_credits <- zb_get_vendor_credits(
    organizations$organization_id
)
}
}
