% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_vendor_credits_details.R
\name{zb_get_vendor_credits_details}
\alias{zb_get_vendor_credits_details}
\title{Get vendor credits details}
\usage{
zb_get_vendor_credits_details(organization_id, vendor_credit_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{vendor_credit_id}{Vector of vendor credit ids}
}
\value{
tibble
}
\description{
Get vendor credits details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/vendorcredits/#get-a-vendor-credit
}
\examples{
\dontrun{
vendor_credits <- zb_get_vendor_credits(organizations$organization_id)
details <- zb_get_vendor_credits_details(
  organization_id = organizations$organization_id[1],
  vendor_credit_id = vendor_credits$vendor_credit_id
)
}
}
