% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_vendor_payments_details.R
\name{zb_get_vendor_payments_details}
\alias{zb_get_vendor_payments_details}
\title{Get vendor payments details}
\usage{
zb_get_vendor_payments_details(organization_id, vendor_payment_id)
}
\arguments{
\item{organization_id}{Organization id}

\item{vendor_payment_id}{Vector of vendor payment ids}
}
\value{
tibble
}
\description{
Get vendor payments details
}
\details{
Uses detail endpoint:
https://www.zoho.com/books/api/v3/vendorpayments/#get-a-vendor-payment
}
\examples{
\dontrun{
vendor_payments <- zb_get_vendor_payments(organizations$organization_id)
details <- zb_get_vendor_payments_details(
  organization_id   = organizations$organization_id[1],
  vendor_payment_id = vendor_payments$vendor_payment_id
)
}
}
