% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzohobooks-package.R
\docType{package}
\name{rzohobooks-package}
\alias{rzohobooks}
\alias{rzohobooks-package}
\title{rzohobooks: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
