% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_credit_notes.R
\name{zb_get_credit_notes}
\alias{zb_get_credit_notes}
\title{Get credit notes}
\usage{
zb_get_credit_notes(organization_id)
}
\arguments{
\item{organization_id}{Organizations ids}
}
\value{
tibble
}
\description{
Get credit notes
}
\details{
See \href{https://www.zoho.com/books/api/v3/credit-notes/#list-all-credit-notes}{api docs}
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
contacts <- zb_get_credit_notes(organizations$organization_id)
}
}
