% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_estimates.R
\name{zb_get_estimates}
\alias{zb_get_estimates}
\title{List all estimates (Quotes)}
\usage{
zb_get_estimates(
  organization_id,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL,
  status = NULL,
  search_text = NULL,
  filter_by = NULL
)
}
\arguments{
\item{organization_id}{Organizations ids}

\item{date_start}{Search estimates by estimate date.}

\item{date_end}{Search estimates by estimate date.}

\item{date_before}{Search estimates by estimate date.}

\item{date_after}{Search estimates by estimate date.}

\item{status}{Search estimates by status.Allowed Values \code{draft}, \code{sent}, \code{invoiced} , \code{accepted}, \code{declined} and \code{expired}.}

\item{search_text}{Search estimates by estimate number or reference or customer name.}

\item{filter_by}{Filter estimates by status.Allowed Values \code{Status.All}, \code{Status.Sent}, \code{Status.Draft}, \code{Status.Invoiced}, \code{Status.Accepted}, \code{Status.Declined} and \code{Status.Expired}.}
}
\description{
List all estimates (Quotes)
}
\details{
For more details see \href{https://www.zoho.com/books/api/v3/estimates/#list-estimates}{API docs}
}
