% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_retainer_invoices.R
\name{zb_get_retainer_invoices}
\alias{zb_get_retainer_invoices}
\title{Get List of Retainer Invoices}
\usage{
zb_get_retainer_invoices(
  organization_id,
  filter_by = NULL,
  date_start = NULL,
  date_end = NULL,
  date_before = NULL,
  date_after = NULL
)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter invoices by any status or payment expected date. Allowed Values: \code{Status.All}, \code{Status.Sent}, \code{Status.Draft}, \code{Status.OverDue}, \code{Status.Paid}, \code{Status.Void}, \code{Status.Unpaid}, \code{Status.PartiallyPaid}, \code{Status.Viewed} and \code{Date.PaymentExpectedDate}.}

\item{date_start}{Start date of period}

\item{date_end}{End date of period}

\item{date_before}{Date before}

\item{date_after}{Date after}
}
\value{
tibble
}
\description{
Get List of Retainer Invoices
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
retainer_invoices <- zb_get_retainer_invoices(
    organizations$organization_id
)
}
}
