% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_get_vendor_payments.R
\name{zb_get_vendor_payments}
\alias{zb_get_vendor_payments}
\title{Get List Vendor Payments}
\usage{
zb_get_vendor_payments(organization_id, filter_by = NULL)
}
\arguments{
\item{organization_id}{Organizations id, you can get list pf organization use \code{zb_get_organizations()}.}

\item{filter_by}{Filter payments by mode. Allowed Values: \code{PaymentMode.All}, \code{PaymentMode.Check}, \code{PaymentMode.Cash}, \code{PaymentMode.BankTransfer}, \code{PaymentMode.Paypal}, \code{PaymentMode.CreditCard}, \code{PaymentMode.GoogleCheckout}, \code{PaymentMode.Credit}, \code{PaymentMode.Authorizenet}, \code{PaymentMode.BankRemittance}, \code{PaymentMode.Payflowpro} and \code{PaymentMode.Others}.}
}
\value{
tibble
}
\description{
Get List Vendor Payments
}
\examples{
\dontrun{
organizations <- zb_get_organizations()
vendor_payments <- zb_get_vendor_payments(
    organizations$organization_id
)
}
}
